[NABI-57] Reorganized Boilerplate
