+++ src/ts/Instructors/ProfileBuilder/test/BasicInfo.unit.test.tsx
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import { Dispatch } from 'redux';
 
 import { BasicInfo } from '../components/BasicInfo/BasicInfo';
-import { Role, UserState } from '../../Registration/model';
+import { Role, UserState } from '../../../Users/model';
 
 describe('BasicInfo', () => {
   let wrapper: any;
