+++ src/ts/Instructors/ProfileBuilder/test/ProfileBuilder.unit.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { ProfileBuilder } from '../';
+import { ProfileBuilder } from '../../';
 import { shallow } from 'enzyme';
-import { Role, UserState } from '../../Registration/model';
+import { Role, UserState } from '../../../Users/model';
 
 describe('ProfileBuilder', () => {
   let wrapper: any;
