+++ src/ts/Users/Registration/test/Registration.test.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import { Registration } from '../';
+import { Registration } from '../..';
 import { shallow } from 'enzyme';
-import { UserState } from '../model';
-import { InstructorState } from '../../ProfileBuilder/model';
+import { UserState } from '../../model';
+import { InstructorState } from '../../../Instructors/model';
 
 describe('Registration', () => {
   let wrapper: any;
