+++ src/ts/Users/index.ts
@@ -1,4 +1,4 @@
-export { default as Registration } from './components/Registration';
+export { default as Registration } from './Registration/components/Registration';
 export * from './actions';
 import * as User from './model';
 export { User };
